requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-docx==1.2.0
pypdf==5.4.0
openpyxl==3.1.5
pytest==8.3.5
httpx==0.28.1
